requirements.txt
# Core LangChain dependencies
langchain==0.1.20
langchain-community==0.0.38
langchain-groq==0.1.3

# HTTP requests
requests==2.31.0

# Environment management  
python-dotenv==1.0.0

# Task scheduling
schedule==1.2.0

# Data validation
pydantic==2.6.4

# Logging and utilities
typing-extensions==4.10.0